requirements.txt
discord.py
opencv-python-headless
pytesseract
numpy
Pillow
python-dotenv
